# Vectorization of find_next_host_delimiter and find_next_host_delimiter_special