--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -145,12 +145,6 @@ ada_really_inline void strip_trailing_spaces_from_opaque_path(
 
 /**
  * @private
- * Reverse the order of the bytes.
- */
-ada_really_inline uint64_t swap_bytes(uint64_t val) noexcept;
-
-/**
- * @private
  * Finds the delimiter of a view in authority state.
  */
 ada_really_inline size_t