We vectorize manually for SSE2 and NEON two functions that relied on SWAR: `find_next_host_delimiter` and `find_next_host_delimiter_special`. To assess the performance effect, we use `benchdata` and the `BasicBench_AdaURL_aggregator_href` measure.

## Apple M2, LLVM 14

Before: speed=484.976M/

After: speed=496.554M/s

## GCC 12, Intel Ice Lake

Before: speed=375.019M/s

After: speed=401.595M/s

## Simpler alternative

If we remove optimizations and switch back to the following (which compiles to a bitset lookup, character-by-character), our speed goes down slightly to 372.901M/s (GCC 12, Intel Ice Lake).

```C++
for (size_t i = location; i < view.size(); i++) {
      if (view[i] == ':' || view[i] == '/' || view[i] == '\\' ||
          view[i] == '?' || view[i] == '[') {
        return i;
      }
    }
    return size_t(view.size());
  }
```

This might compile to...

```asm
        movabs  rcx, 17592186112001
        mov     rax, rdx
.L2:
        cmp     rax, rdi
        jnb     .L7
        mov     dl, BYTE PTR [rsi+rax]
        sub     edx, 47
        cmp     dl, 44
        ja      .L3
        bt      rcx, rdx
        jc      .L1
.L3:
        inc     rax
        jmp     .L2
```

... which is likely highly competitive with a character-by-character table lookup (because of the latency of a table lookup).

Fixes https://github.com/ada-url/ada/issues/547

cc @the-moisrex