@the-moisrex 2023-11-01
Seems like a simple lookup table is faster than the current `main` branch, and this PR is faster than a simplified version; but proposed benchmark does a ton of allocations and what not, and I saw a lot of noises in this benchmark (possibly) due to those.

| Version                   | Mean                            | Median                         |
| --------------------------| ------------------------------- | ------------------------------ |
| Simplified               | `speed=289.901M/s`    | `speed=290.272M/s`   |
| Main                        | `speed=267.224M/s`  | `speed=260.651M/s`   |
| Vectorized(SSE2)   | `speed=307.408M/s`  | `speed=307.465M/s`   |

```bash
./benchmarks/benchdata --benchmark_filter=BasicBench_AdaURL_aggregator_href --benchmark_color=true --benchmark_repetitions=30
```


Simplified:
```c++
// : / [ \\ ?
static constexpr std::array<bool, 128> special_host_delimiters{
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, true,  false, false, false, false, false, false, false,
    false, false, false, true,  false, false, false, false, true,  false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, true,  true,  false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false,
};

// : / [ ?
static constexpr std::array<bool, 128> host_delimiters{
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, true,  false, false, false, false, false, false, false,
    false, false, false, true,  false, false, false, false, true,  false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, true,  false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false, false, false, false, false,
    false, false, false, false, false, false, false,
};

// starting at index location, this finds the next location of a character
// :, /, \\, ? or [. If none is found, view.size() is returned.
// For use within get_host_delimiter_location.
// ['0x3a', '0x2f', '0x5c', '0x3f', '0x5b']
ada_really_inline size_t find_next_host_delimiter_special(
    std::string_view view, size_t location) noexcept {
  auto const str = view.substr(location);
  for (auto pos = str.begin(); pos != str.end(); ++pos) {
    if (special_host_delimiters[*pos]) {
      return pos - str.begin();
    }
  }
  return size_t(view.size());
}

// starting at index location, this finds the next location of a character
// :, /, ? or [. If none is found, view.size() is returned.
// For use within get_host_delimiter_location.

ada_really_inline size_t find_next_host_delimiter(std::string_view view,
                                                  size_t location) noexcept {
  auto const str = view.substr(location);
  for (auto pos = str.begin(); pos != str.end(); ++pos) {
    if (host_delimiters[*pos]) {
      return pos - str.begin();
    }
  }
  return size_t(view.size());
}

```