@the-moisrex 2023-11-01
Let me know if you want the simplified version as a separate PR.