@lemire 2023-11-01
@the-moisrex Please see https://github.com/ada-url/ada/pull/549