@lemire 2023-11-01
@the-moisrex I have updated this PR. We fallback on your proposal when SIMD is not available. I added credit to you in the comments.