@the-moisrex 2023-11-01
> > I haven't tested the std::bitset way of doing this here, but that's generally slower than look up table.
> 
> By bitset, I did not mean `std::bitset`. The compiler will use a single 64-bit word as a bitset.

The approach in `character_sets-inl.h` is almost the same way that `std::bitset` does it, that's why I mentioned it.