@lemire 2023-11-01
> The approach in character_sets-inl.h is almost the same way that std::bitset does it, that's why I mentioned it.

Yep. The percent coding work can be accelerated.

We should vectorize it, at least on some systems. We have a PR on this front.